+++ lib/mongoid/contextual/atomic.rb
@@ -6,84 +6,76 @@ module Mongoid
       # Execute an atomic $addToSet on the matching documents.
       #
       # @example Add the value to the set.
-      #   context.add_to_set(:members, "Dave")
+      #   context.add_to_set(members: "Dave", genres: "Electro")
       #
-      # @param [ String, Symbol ] field The name of the field to add to.
-      # @param [ Object ] value The single value to add.
+      # @param [ Hash ] adds The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def add_to_set(field, value)
-        query.update_all("$addToSet" => { database_field_name(field) => value })
+      def add_to_set(adds)
+        query.update_all("$addToSet" => collect_operations(adds))
       end
 
       # Perform an atomic $bit operation on the matching documents.
       #
       # @example Perform the bitwise op.
-      #   context.bit(:likes, { and: 14, or: 4 })
+      #   context.bit(likes: { and: 14, or: 4 })
       #
-      # @param [ String, Symbol ] field The name of the field to operate on.
-      # @param [ Hash ] value The bitwise operations to perform. Keys may be
-      #   "and" or "or" and must have numeric values.
+      # @param [ Hash ] bits The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def bit(field, value)
-        query.update_all("$bit" => { database_field_name(field) => value })
+      def bit(bits)
+        query.update_all("$bit" => collect_operations(bits))
       end
 
       # Perform an atomic $inc operation on the matching documents.
       #
       # @example Perform the atomic increment.
-      #   context.inc(:likes, 10)
+      #   context.inc(likes: 10)
       #
-      # @param [ String, Symbol ] field The field to increment.
-      # @param [ Integer ] value The amount to increment by.
+      # @param [ Hash ] incs The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def inc(field, value)
-        query.update_all("$inc" => { database_field_name(field) => value })
+      def inc(incs)
+        query.update_all("$inc" => collect_operations(incs))
       end
 
       # Perform an atomic $pop operation on the matching documents.
       #
       # @example Pop the first value on the matches.
-      #   context.pop(:members, -1)
+      #   context.pop(members: -1)
       #
       # @example Pop the last value on the matches.
-      #   context.pop(:members, 1)
+      #   context.pop(members: 1)
       #
-      # @param [ String, Symbol ] field The name of the array field to pop
-      #   from.
-      # @param [ Integer ] value 1 to pop from the end, -1 to pop from the
-      #   front.
+      # @param [ Hash ] pops The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def pop(field, value)
-        query.update_all("$pop" => { database_field_name(field) => value })
+      def pop(pops)
+        query.update_all("$pop" => collect_operations(pops))
       end
 
       # Perform an atomic $pull operation on the matching documents.
       #
       # @example Pull the value from the matches.
-      #   context.pull(:members, "Dave")
+      #   context.pull(members: "Dave")
       #
       # @note Expression pulling is not yet supported.
       #
-      # @param [ String, Symbol ] field The field to pull from.
-      # @param [ Object ] value The single value to pull.
+      # @param [ Hash ] pulls The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def pull(field, value)
-        query.update_all("$pull" => { database_field_name(field) => value })
+      def pull(pulls)
+        query.update_all("$pull" => collect_operations(pulls))
       end
 
       # Perform an atomic $pullAll operation on the matching documents.
@@ -91,74 +83,73 @@ module Mongoid
       # @example Pull all the matching values from the matches.
       #   context.pull_all(:members, [ "Alan", "Vince" ])
       #
-      # @param [ String, Symbol ] field The field to pull from.
-      # @param [ Array<Object> ] values The values to pull.
+      # @param [ Hash ] pulls The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def pull_all(field, values)
-        query.update_all("$pullAll" => { database_field_name(field) => values })
+      def pull_all(pulls)
+        query.update_all("$pullAll" => collect_operations(pulls))
       end
 
       # Perform an atomic $push operation on the matching documents.
       #
       # @example Push the value to the matching docs.
-      #   context.push(:members, "Alan")
+      #   context.push(members: "Alan")
       #
-      # @param [ String, Symbol ] field The field to push to.
-      # @param [ Object ] value The value to push.
+      # @param [ Hash ] pushes The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def push(field, value)
-        query.update_all("$push" => { database_field_name(field) => value })
+      def push(pushes)
+        query.update_all("$push" => collect_operations(pushes))
       end
 
       # Perform an atomic $pushAll operation on the matching documents.
       #
       # @example Push the values to the matching docs.
-      #   context.push(:members, [ "Alan", "Fletch" ])
+      #   context.push(members: [ "Alan", "Fletch" ])
       #
-      # @param [ String, Symbol ] field The field to push to.
-      # @param [ Array<Object> ] values The values to push.
+      # @param [ Hash ] pushes The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def push_all(field, values)
-        query.update_all("$pushAll" => { database_field_name(field) => values })
+      def push_all(pushes)
+        query.update_all("$pushAll" => collect_operations(pushes))
       end
 
       # Perform an atomic $rename of fields on the matching documents.
       #
       # @example Rename the fields on the matching documents.
-      #   context.rename(:members, :artists)
+      #   context.rename(members: :artists)
       #
-      # @param [ String, Symbol ] old_name The old field name.
-      # @param [ String, Symbol ] new_name The new field name.
+      # @param [ Hash ] renames The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def rename(old_name, new_name)
-        query.update_all("$rename" => { database_field_name(old_name) => new_name.to_s })
+      def rename(renames)
+        operations = renames.inject({}) do |ops, (old_name, new_name)|
+          ops[old_name] = new_name.to_s
+          ops
+        end
+        query.update_all("$rename" => collect_operations(operations))
       end
 
       # Perform an atomic $set of fields on the matching documents.
       #
       # @example Set the field value on the matches.
-      #   context.set(:name, "Depeche Mode")
+      #   context.set(name: "Depeche Mode")
       #
-      # @param [ String, Symbol ] field The name of the field.
-      # @param [ Object ] value The value to set.
+      # @param [ Hash ] sets The operations.
       #
       # @return [ nil ] Nil.
       #
       # @since 3.0.0
-      def set(field, value)
-        query.update_all("$set" => { database_field_name(field) => value })
+      def set(sets)
+        query.update_all("$set" => collect_operations(sets))
       end
 
       # Perform an atomic $unset of a field on the matching documents.
@@ -174,6 +165,15 @@ module Mongoid
       def unset(*args)
         fields = args.__find_args__.collect { |f| [database_field_name(f), true] }
         query.update_all("$unset" => Hash[fields])
+      end
+
+      private
+
+      def collect_operations(ops)
+        ops.inject({}) do |operations, (field, value)|
+          operations[database_field_name(field)] = value.mongoize
+          operations
+        end
       end
     end
   end