+++ lib/mongoid/relations/counter_cache.rb
@@ -40,9 +40,7 @@ module Mongoid
         #
         # @since 3.1.0
         def update_counters(id, counters)
-          counters.map do |key, value|
-            where(:_id => id).inc(key, value)
-          end
+          where(:_id => id).inc(counters)
         end
 
         # Increment the counter name from the entries that match the