+++ lib/mongoid/relations/synchronization.rb
@@ -61,7 +61,7 @@ module Mongoid
       def remove_inverse_keys(meta)
         foreign_keys = send(meta.foreign_key)
         unless foreign_keys.nil? || foreign_keys.empty?
-          meta.criteria(foreign_keys, self.class).pull(meta.inverse_foreign_key, id)
+          meta.criteria(foreign_keys, self.class).pull(meta.inverse_foreign_key => id)
         end
       end
 
@@ -92,10 +92,10 @@ module Mongoid
           end
 
           unless adds.empty?
-            meta.criteria(adds, self.class).without_options.add_to_set(meta.inverse_foreign_key, id)
+            meta.criteria(adds, self.class).without_options.add_to_set(meta.inverse_foreign_key => id)
           end
           unless subs.empty?
-            meta.criteria(subs, self.class).without_options.pull(meta.inverse_foreign_key, id)
+            meta.criteria(subs, self.class).without_options.pull(meta.inverse_foreign_key => id)
           end
         end
       end